# docs/swagger.yaml
definitions:
  controllers.CityInput:
    properties:
      name:
        type: string
    type: object
...
host: www.tripwise.my.id
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Documentation TripWise Capstone Project - Team 3
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TRIPWISE API
  version: "1.0"
paths:
...
  /city:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all cities from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all cities
      tags:
      - Cities
    post:
      consumes:
      - application/json
      description: Create a new city in the database
      parameters:
      - description: City Name
        in: body
        name: city
        required: true
        schema:
          $ref: '#/definitions/controllers.CityInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new city
      tags:
      - Cities
  /dashboard/count-data:
    get:
      description: Retrieve a summary of users, destinations, video content, and destination
        categories for the dashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Fetch dashboard data summary
      tags:
      - Dashboard
  /dashboard/graphic:
    get:
      description: Retrieve the number of user registrations grouped by month
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Fetch monthly user registration data
      tags:
      - Dashboard
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"